+++ app/Http/Controllers/Api/CategoryController.php
@@ -16,6 +16,7 @@ class CategoryController extends Controller
      */
     public function index(Company $company)
     {
+        $this->authorize('viewAny', Category::class);
         $categories = $company->categories;
         return response()->json([
             'message' => 'Categories retrieved successfully',
@@ -28,6 +29,7 @@ class CategoryController extends Controller
      */
     public function store(StoreCategoryRequest $request, Company $company)
     {
+        $this->authorize('create', Category::class);
         $category = $company->categories()->create($request->validated());
 
         return response()->json([
@@ -41,6 +43,7 @@ class CategoryController extends Controller
      */
     public function show(Company $company, Category $category)
     {
+        $this->authorize('view', $category);
         return response()->json([
             'message' => 'Category retrieved successfully',
             'data' => $category
@@ -52,6 +55,7 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Company $company, Category $category)
     {
+        $this->authorize('update', $category);
         $category->update($request->validated());
 
         return response()->json([
@@ -65,6 +69,7 @@ class CategoryController extends Controller
      */
     public function destroy(Company $company, Category $category)
     {
+        $this->authorize('delete', $category);
         $category->delete();
 
         return response()->json([
@@ -78,6 +83,7 @@ class CategoryController extends Controller
     public function restore(Company $company, $id)
     {
         $category = $company->categories()->onlyTrashed()->findOrFail($id);
+        $this->authorize('restore', $category);
         $category->restore();
 
         return response()->json([
@@ -92,6 +98,7 @@ class CategoryController extends Controller
     public function forceDelete(Company $company, $id)
     {
         $category = $company->categories()->onlyTrashed()->findOrFail($id);
+        $this->authorize('forceDelete', $category);
         $category->forceDelete();
 
         return response()->json([
