+++ app/Http/Controllers/Api/OrderController.php
@@ -16,6 +16,7 @@ class OrderController extends Controller
      */
     public function index(Company $company)
     {
+        $this->authorize('viewAny', Order::class);
         $orders = $company->orders;
         return response()->json([
             'message' => 'Orders retrieved successfully',
@@ -28,6 +29,7 @@ class OrderController extends Controller
      */
     public function store(StoreOrderRequest $request, Company $company)
     {
+        $this->authorize('create', Order::class);
         $order = $company->orders()->create($request->validated());
 
         return response()->json([
@@ -41,6 +43,7 @@ class OrderController extends Controller
      */
     public function show(Company $company, Order $order)
     {
+        $this->authorize('view', $order);
         return response()->json([
             'message' => 'Order retrieved successfully',
             'data' => $order
@@ -52,6 +55,7 @@ class OrderController extends Controller
      */
     public function update(UpdateOrderRequest $request, Company $company, Order $order)
     {
+        $this->authorize('update', $order);
         $order->update($request->validated());
 
         return response()->json([
@@ -65,6 +69,7 @@ class OrderController extends Controller
      */
     public function destroy(Company $company, Order $order)
     {
+        $this->authorize('delete', $order);
         $order->delete();
 
         return response()->json([
@@ -78,6 +83,7 @@ class OrderController extends Controller
     public function restore(Company $company, $id)
     {
         $order = $company->orders()->onlyTrashed()->findOrFail($id);
+        $this->authorize('restore', $order);
         $order->restore();
 
         return response()->json([
@@ -92,6 +98,7 @@ class OrderController extends Controller
     public function forceDelete(Company $company, $id)
     {
         $order = $company->orders()->onlyTrashed()->findOrFail($id);
+        $this->authorize('forceDelete', $order);
         $order->forceDelete();
 
         return response()->json([
